+++ folsom/src/main/java/com/spotify/folsom/MemcacheClientBuilder.java
@@ -556,7 +556,7 @@ public class MemcacheClientBuilder<V> {
     return this;
   }
 
-  public MemcacheClientBuilder<V> withSSLEngineFactory(SSLEngineFactory sslEngineFactory) {
+  public MemcacheClientBuilder<V> withSSLEngineFactory(final SSLEngineFactory sslEngineFactory) {
     this.sslEngineFactory = sslEngineFactory;
     return this;
   }
