+++ folsom/src/main/java/com/spotify/folsom/client/tls/DefaultSSLEngineFactory.java
@@ -6,18 +6,27 @@ import java.security.NoSuchAlgorithmException;
 
 public class DefaultSSLEngineFactory implements SSLEngineFactory {
   final SSLContext sslContext;
+  final boolean reuseSession;
 
-  public DefaultSSLEngineFactory() throws NoSuchAlgorithmException {
-    this(SSLContext.getDefault());
+  public DefaultSSLEngineFactory(final boolean reuseSession) throws NoSuchAlgorithmException {
+    this(SSLContext.getDefault(), reuseSession);
   }
 
-  public DefaultSSLEngineFactory(SSLContext sslContext) {
+  public DefaultSSLEngineFactory(final SSLContext sslContext, final boolean reuseSession) {
     this.sslContext = sslContext;
+    this.reuseSession = reuseSession;
   }
 
   @Override
   public SSLEngine createSSLEngine(String hostname, int port) {
-    SSLEngine sslEngine = sslContext.createSSLEngine(hostname, port);
+    final SSLEngine sslEngine;
+
+    if(reuseSession) {
+      sslEngine = sslContext.createSSLEngine(hostname, port);
+    } else {
+      sslEngine = sslContext.createSSLEngine();
+    }
+
     sslEngine.setUseClientMode(true);
     return sslEngine;
   }
